Bird-Classification/requirements.txt
Pillow==9.3.0
onnxruntime==1.17.3
torch==2.3.1
torchvision==0.18.1
numpy==1.24.2
fastapi==0.111.0
uvicorn==0.29.0